dataset/langchain-chatbot-main/requirements.txt
accelerate==0.29.3
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
auto-gptq==0.7.1
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.7.8
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.19.2
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
duckdb==0.10.2
effdet==0.4.1
emoji==2.12.1
eval-type-backport==0.2.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.85.1
filelock==3.13.3
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.2.0
gekko==1.1.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
graphlib-backport==1.1.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.4.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.14
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.38
layoutparser==0.3.4
lorem-text==2.1
lxml==5.2.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.5
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
mteb==1.12.23
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.15.1
openai==1.16.0
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
optimum==1.19.0
ordered-set==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
peft==0.10.0
pikepdf==8.15.1
pillow==10.3.0
polars==0.20.31
portalocker==2.8.2
posthog==3.5.0
protobuf==5.26.1
psutil==5.9.8
pulsar-client==3.5.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.3
pydantic-core==2.18.4
pygments==2.18.0
pyparsing==3.1.2
pypdf==4.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject-hooks==1.1.0
pysbd==0.3.4
pysqlite3-binary==0.5.2.post3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytrec-eval-terrier==0.5.6
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.8
rapidfuzz==3.9.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rouge==1.0.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==2.1.0
starlette==0.20.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
timm==1.0.3
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.2
torchvision==0.18.0
tqdm==4.66.2
transformers==4.39.2
triton==2.2.0
typer==0.12.3
types-requests==2.31.0.20240402
typing-extensions==4.10.0
typing-inspect==0.9.0
tzdata==2024.1
unstructured==0.11.8
unstructured-client==0.22.0
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
